# storage/api-docs/api-docs.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "TasteAsia RESTful API",
        "description": "RESTful API Documentation for TasteAsia. Explore our endpoints and discover how to interact with our delicious data!",
        "contact": {
            "email": "gabrielaswinta@cosrsivalab.space"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://tasteasia.test",
            "description": "TasteAsia API Server"
        },
        {
            "url": "https://tasteasia.test/api/v1",
            "description": "Local Development Server"
        },
        {
            "url": "${STAGING_URL}/api/v1",
            "description": "STAGING API Server"
        },
        {
            "url": "https://tasteasia.test/api/v1",
            "description": "PRODUCTION API Server"
        }
    ],
    "paths": {
...
        "/m/location/create": {
            "post": {
                "tags": [
                    "Location"
                ],
                "summary": "Add new location",
                "description": "Add new location",
                "operationId": "face1662e4f548ee18e79092758d3502",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the location",
                                        "type": "string",
                                        "example": "SIT - Singapore Institute of Technology"
                                    },
                                    "address": {
                                        "description": "address of the location",
                                        "type": "string",
                                        "example": "11 New Punggol Road"
                                    },
                                    "url": {
                                        "description": "location URL",
                                        "type": "string",
                                        "example": "https://maps.app.goo.gl/K1oMjsLwU2p2ZJo7A"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Indicates if the transaction was successful",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "location": {
                                            "description": "location data",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/m/location/all": {
            "get": {
                "tags": [
                    "Location"
                ],
                "summary": "Get location list",
                "description": "Retrieves all location",
                "operationId": "e6b8bc834f3169b94eca8374f0f9f695",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "The ID of the location",
                                            "type": "integer"
                                        },
                                        "name": {
                                            "description": "Location name",
                                            "type": "string"
                                        },
                                        "address": {
                                            "description": "Location Address",
                                            "type": "string"
                                        },
                                        "country": {
                                            "description": "Location country",
                                            "type": "string"
                                        },
                                        "postal_code": {
                                            "description": "Location Postal Code",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    },
                    {
                        "Access Token": []
                    }
                ]
            }
        },
        "/m/location/toggle-status": {
            "post": {
                "tags": [
                    "Location"
                ],
                "summary": "Toggle location status between active and inactive",
                "description": "Toggle location status",
                "operationId": "f32dbb93340dfa4fd74744abfdd05eec",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "location_id": {
                                        "description": "Location ID",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "status": {
                                        "description": "Location status",
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Indicates if the transaction was successful",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "description": "response message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "Bearer Token": {
                "type": "http",
                "description": "Enter token in format 'Bearer {token}'. Token is a 64-character random string.",
                "scheme": "bearer"
            },
            "Access Token": {
                "type": "apiKey",
                "description": "Sanctum Access Token (a random string)",
                "name": "X-Taste-Asia-Access-Token",
                "in": "header"
            },
            "Refresh Token": {
                "type": "apiKey",
                "description": "Sanctum Refresh Token (a random string)",
                "name": "X-Taste-Asia-Refresh-Token",
                "in": "header"
            }
        }
    },
...
}